- Made Interbase compile under Win32, cleaned warnings @- Added Win32 project for the Interbase module (Zeev)

+++ ext/interbase/php_interbase.h
@@ -96,7 +96,7 @@ ZEND_END_MODULE_GLOBALS(ibase)
 typedef struct {
 	isc_tr_handle trans[IBASE_TRANS_ON_LINK];
 	isc_db_handle link;
-	int dialect;
+	unsigned short dialect;
 } ibase_db_link;
 
 typedef struct {
@@ -126,7 +126,7 @@ typedef struct {
 	XSQLDA *in_sqlda, *out_sqlda;
 	ibase_array *in_array, *out_array;
 	int in_array_cnt, out_array_cnt;
-	int dialect;
+	unsigned short dialect;
 	int cursor_open;
 } ibase_query;
